# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /passkey/register/options:
    post:
      operationId: passkeyRegisterOptions
      tags: [passkey]
      summary: Generate WebAuthn registration options
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasskeyRegisterOptionsBody"
      responses:
        "200":
          description: Options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasskeyRegisterOptionsResponse"
  /passkey/register/verify:
    post:
      operationId: passkeyRegisterVerify
      tags: [passkey]
      summary: Verify WebAuthn registration response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasskeyRegisterVerifyBody"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasskeyRegisterVerifyResponse"
  /passkey/authenticate/options:
    post:
      operationId: passkeyAuthenticateOptions
      tags: [passkey]
      summary: Generate WebAuthn authentication options
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasskeyAuthenticateOptionsBody"
      responses:
        "200":
          description: Options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasskeyAuthenticateOptionsResponse"
  /passkey/authenticate/verify:
    post:
      operationId: passkeyAuthenticateVerify
      tags: [passkey]
      summary: Verify WebAuthn authentication response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasskeyAuthenticateVerifyBody"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasskeyAuthenticateVerifyResponse"
  /passkey/refresh:
    post:
      operationId: passkeyRefresh
      tags: [passkey]
      summary: Refresh session token
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionTokenResponse"
  /passkey/logout:
    post:
      operationId: passkeyLogout
      tags: [passkey]
      summary: Logout (revoke session token)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
      required: [success]
...
    PasskeyRegisterOptionsBody:
      type: object
      properties:
        username: { type: string, minLength: 1, maxLength: 200 }
      required: [username]
    PasskeyRegisterOptionsResponse:
      type: object
      additionalProperties: true
      description: Options object returned by simplewebauthn for navigator.credentials.create
    PasskeyRegisterVerifyBody:
      type: object
      properties:
        username: { type: string, minLength: 1, maxLength: 200 }
        response:
          type: object
          additionalProperties: true
      required: [username, response]
    PasskeyRegisterVerifyResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        userId: { type: string }
        user:
          $ref: "#/components/schemas/UserPublic"
        token: { type: string }
        expiresAt: { type: integer, format: int64 }
      required: [success, userId, token, expiresAt, user]
    PasskeyAuthenticateOptionsBody:
      type: object
      properties:
        username: { type: string, minLength: 1, maxLength: 200 }
    PasskeyAuthenticateOptionsResponse:
      type: object
      additionalProperties: true
      description: Options object returned by simplewebauthn for navigator.credentials.get
    PasskeyAuthenticateVerifyBody:
      type: object
      properties:
        response:
          type: object
          additionalProperties: true
      required: [response]
    PasskeyAuthenticateVerifyResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        user:
          $ref: "#/components/schemas/UserPublic"
        token: { type: string }
        expiresAt: { type: integer, format: int64 }
      required: [success, user, token, expiresAt]
    SessionTokenResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        token: { type: string }
        expiresAt: { type: integer, format: int64 }
      required: [success, token, expiresAt]
    UserPublic:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        email:
          oneOf:
            - { type: string }
            - { type: "null" }
      required: [id, username]
...